pyproject.toml
[project]
name = "follow-pay"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.2.1,<8.0.0)",
    "stripe (>=14.4.0,<15.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "drf-yasg (>=1.21.15,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
test = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.git          # в корне проекта
  )/
)
'''

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = """
--cov=users
--cov=creators
--cov=posts
--cov=subscriptions
--cov=payments
--cov=web
--cov-report=term-missing
--cov-fail-under=80
"""

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests.py",
    "tests/*",
    "*/admin.py",
    "*/apps.py",
    "*/urls.py",
    "config/settings.py",
    "config/wsgi.py",
    "config/asgi.py",
    "config/__init__.py",
    "manage.py",
    "web/views.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]